se desacopla componente de App.js
